Ensure native extensions are compiled for linux
If you are not working in linux, we need to cross-compile any ruby gems
that have native extensions.

This step is skipped if none of your gems have native extensions, or if
you are working on a linux system (so we can assume the gems are already
installed and compiled).