Fix native extension builds for linux
When developing on linux, the native extensions may be built to
`extensions/x86_64-linux/2.5.0` - but lambda expects them to be packaged
to `extensions/x86_64-linux/2.5.0-static`.
For now, we will create a symlink so the files are available at `2.5.0-static`
if that directory does not already exist.
A future enhancement should detect the actual extension directory, instead
of assuming it is always `2.5.0`.